Improved mining ai
