+++ src/main/java/com/fi0x/deepmagic/entities/ai/EntityAIDigRandom.java
@@ -1,58 +1,119 @@
 package com.fi0x.deepmagic.entities.ai;
 
 import com.fi0x.deepmagic.entities.EntityDwarf;
-import net.minecraft.entity.ai.EntityAIWanderAvoidWater;
+import com.fi0x.deepmagic.init.ModBlocks;
+import net.minecraft.block.Block;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.EntityCreature;
+import net.minecraft.entity.ai.EntityAIBase;
+import net.minecraft.init.Blocks;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
-public class EntityAIDigRandom extends EntityAIWanderAvoidWater
-{
-    World world;
+import java.util.ArrayList;
 
-    public EntityAIDigRandom(EntityDwarf entity, World world)
+public class EntityAIDigRandom extends EntityAIBase
+{
+    protected EntityDwarf creature;
+    private BlockPos destination;
+    private final World world;
+    private ArrayList<IBlockState> mineableBlocks;
+
+    public EntityAIDigRandom(EntityDwarf creature, World world)
     {
-        super(entity, 1, 30);
+        this.creature = creature;
         this.setMutexBits(1);
         this.world = world;
+
+        mineableBlocks = new ArrayList<>();
+        mineableBlocks.add(Blocks.STONE.getDefaultState());
+        mineableBlocks.add(Blocks.DIRT.getDefaultState());
+        mineableBlocks.add(Blocks.AIR.getDefaultState());
+        mineableBlocks.add(Blocks.QUARTZ_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.COAL_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.IRON_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.GOLD_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.REDSTONE_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.LAPIS_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.DIAMOND_ORE.getDefaultState());
+        mineableBlocks.add(Blocks.EMERALD_ORE.getDefaultState());
+        mineableBlocks.add(ModBlocks.DEEP_CRYSTAL_ORE.getDefaultState());
     }
 
     @Override
     public boolean shouldExecute()
     {
-        return super.shouldExecute();
+        return !creature.isMining && creature.posY < 50;
     }
 
     @Override
     public void startExecuting()
     {
-        super.startExecuting();
+//        this.creature.isMining = true;
+//        destination = getRandomDestination();
+//
+//        ArrayList<BlockPos> miningBlocks = getMiningPath(creature.getPosition(), destination);
+//        if(isMineablePath(world, miningBlocks)) mineBlocks(world, miningBlocks, creature);
+//        else creature.isMining = false;
     }
 
     @Override
     public boolean shouldContinueExecuting()
     {
-        if(entity.getNavigator().noPath())
+        if(creature.getPosition() == destination) creature.isMining = false;
+        return creature.isMining;
+    }
+
+    private BlockPos getRandomDestination()
+    {
+        return new BlockPos(creature.posX + (Math.random() * 20) - 10, creature.posY, creature.posZ + (Math.random() * 20) - 10);
+    }
+
+    public ArrayList<BlockPos> getMiningPath(BlockPos currentPosition, BlockPos destination)
+    {
+        ArrayList<BlockPos> blocks = new ArrayList<>();
+        int xIncrease = 1;
+        int zIncrease = 1;
+        if(destination.getX() > currentPosition.getX()) xIncrease = -1;
+        if(destination.getZ() > currentPosition.getZ()) zIncrease = -1;
+
+        BlockPos checkPos = currentPosition;
+        while(checkPos.getX() != destination.getX())
         {
-            BlockPos pos = getRandomBlock();
-            world.getBlockState(pos).getBlock().dropBlockAsItem(world, pos, world.getBlockState(pos).getBlock().getDefaultState(), 1);
-            world.setBlockToAir(pos);
-            pos = pos.add(0, (Math.random() * 3) - 1, 0);
-            world.getBlockState(pos).getBlock().dropBlockAsItem(world, pos, world.getBlockState(pos).getBlock().getDefaultState(), 1);
-            world.setBlockToAir(pos);
-            return false;
+            checkPos = checkPos.add(xIncrease, 0, 0);
+            blocks.add(checkPos.add(0, 1, 0));
+            blocks.add(checkPos);
+        }
+        while(checkPos.getZ() != destination.getZ())
+        {
+            checkPos = checkPos.add(0, 0, zIncrease);
+            blocks.add(checkPos.add(0, 1, 0));
+            blocks.add(checkPos);
+        }
+
+        return blocks;
+    }
+
+    public boolean isMineablePath(World world, ArrayList<BlockPos> blocks)
+    {
+        for (BlockPos block : blocks)
+        {
+            if (!mineableBlocks.contains(world.getBlockState(block))) return false;
         }
         return true;
     }
 
-    private BlockPos getRandomBlock()
+    public void mineBlocks(World world, ArrayList<BlockPos> miningBlocks, EntityCreature creature)
     {
-        int x = 0;
-        int z = 0;
-        while(x == 0 && z == 0)
+        while(!miningBlocks.isEmpty())
         {
-            x = (int) (Math.random() * 3) - 1;
-            if(x == 0) z = (int) (Math.random() * 3) - 1;
+            BlockPos currentPosition = miningBlocks.get(0);
+            Block currentBlock = world.getBlockState(currentPosition).getBlock();
+            currentBlock.dropBlockAsItem(world, miningBlocks.get(0), currentBlock.getDefaultState(), 2);
+            world.setBlockToAir(currentPosition);
+            creature.getNavigator().tryMoveToXYZ(currentPosition.getX(), currentPosition.getY(), currentPosition.getZ(), 1);
+
+            miningBlocks.remove(0);
         }
-        return entity.getPosition().add(x, 0, z);
     }
 }
