+++ src/main/java/com/fi0x/deepmagic/entities/ai/EntityAIMining.java
@@ -1,119 +1,100 @@
 package com.fi0x.deepmagic.entities.ai;
 
 import com.fi0x.deepmagic.entities.EntityDwarf;
-import com.fi0x.deepmagic.init.ModBlocks;
-import net.minecraft.block.Block;
-import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.EntityCreature;
 import net.minecraft.entity.ai.EntityAIBase;
-import net.minecraft.init.Blocks;
+import net.minecraft.entity.ai.RandomPositionGenerator;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
-
-import java.util.ArrayList;
 
 public class EntityAIMining extends EntityAIBase
 {
-    protected final EntityDwarf creature;
-    private BlockPos destination;
-    private final World world;
-    private final ArrayList<IBlockState> mineableBlocks;
+    protected final int executionChance;
+    protected final EntityCreature entity;
+    protected final World world;
+    protected double x;
+    protected double y;
+    protected double z;
+    protected final double speed;
+    protected final float probability;
 
-    public EntityAIMining(EntityDwarf creature, World world)
+    public EntityAIMining(EntityDwarf entity)
     {
-        this.creature = creature;
-        this.setMutexBits(1);
-        this.world = world;
-
-        mineableBlocks = new ArrayList<>();
-        mineableBlocks.add(Blocks.STONE.getDefaultState());
-        mineableBlocks.add(Blocks.DIRT.getDefaultState());
-        mineableBlocks.add(Blocks.AIR.getDefaultState());
-        mineableBlocks.add(Blocks.QUARTZ_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.COAL_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.IRON_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.GOLD_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.REDSTONE_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.LAPIS_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.DIAMOND_ORE.getDefaultState());
-        mineableBlocks.add(Blocks.EMERALD_ORE.getDefaultState());
-        mineableBlocks.add(ModBlocks.DEEP_CRYSTAL_ORE.getDefaultState());
+        this(entity, 1);
+    }
+    public EntityAIMining(EntityDwarf entity, int speed)
+    {
+        this(entity, speed, 120);
+    }
+    public EntityAIMining(EntityDwarf entity, int speed, int executionChance)
+    {
+        this.setMutexBits(3);
+        this.entity = entity;
+        this.world = entity.world;
+        this.probability = 0.001F;
+        this.speed = speed;
+        this.executionChance = executionChance;
     }
 
     @Override
     public boolean shouldExecute()
     {
-        return !creature.isMining && creature.posY < 50;
+        if (this.entity.getIdleTime() >= 100) return false;
+        if (this.entity.getRNG().nextInt(this.executionChance) != 0) return false;
+
+        Vec3d vec3d = this.getPosition();
+
+        if (vec3d == null) return false;
+        else
+        {
+            this.x = vec3d.x;
+            this.y = vec3d.y;
+            this.z = vec3d.z;
+            return true;
+        }
     }
 
     @Override
     public void startExecuting()
     {
-//        this.creature.isMining = true;
-//        destination = getRandomDestination();
-//
-//        ArrayList<BlockPos> miningBlocks = getMiningPath(creature.getPosition(), destination);
-//        if(isMineablePath(world, miningBlocks)) mineBlocks(world, miningBlocks, creature);
-//        else creature.isMining = false;
+        this.entity.getNavigator().tryMoveToXYZ(this.x, this.y, this.z, this.speed);
     }
 
     @Override
     public boolean shouldContinueExecuting()
     {
-        if(creature.getPosition() == destination) creature.isMining = false;
-        return creature.isMining;
-    }
-
-    private BlockPos getRandomDestination()
-    {
-        return new BlockPos(creature.posX + (Math.random() * 20) - 10, creature.posY, creature.posZ + (Math.random() * 20) - 10);
-    }
-
-    public ArrayList<BlockPos> getMiningPath(BlockPos currentPosition, BlockPos destination)
-    {
-        ArrayList<BlockPos> blocks = new ArrayList<>();
-        int xIncrease = 1;
-        int zIncrease = 1;
-        if(destination.getX() > currentPosition.getX()) xIncrease = -1;
-        if(destination.getZ() > currentPosition.getZ()) zIncrease = -1;
-
-        BlockPos checkPos = currentPosition;
-        while(checkPos.getX() != destination.getX())
+        if(entity.getNavigator().noPath())
         {
-            checkPos = checkPos.add(xIncrease, 0, 0);
-            blocks.add(checkPos.add(0, 1, 0));
-            blocks.add(checkPos);
-        }
-        while(checkPos.getZ() != destination.getZ())
-        {
-            checkPos = checkPos.add(0, 0, zIncrease);
-            blocks.add(checkPos.add(0, 1, 0));
-            blocks.add(checkPos);
-        }
-
-        return blocks;
-    }
-
-    public boolean isMineablePath(World world, ArrayList<BlockPos> blocks)
-    {
-        for (BlockPos block : blocks)
-        {
-            if (!mineableBlocks.contains(world.getBlockState(block))) return false;
+            BlockPos pos = getRandomBlock();
+            world.getBlockState(pos).getBlock().dropBlockAsItem(world, pos, world.getBlockState(pos).getBlock().getDefaultState(), 1);
+            world.setBlockToAir(pos);
+            pos = pos.add(0, (Math.random() * 3) - 1, 0);
+            world.getBlockState(pos).getBlock().dropBlockAsItem(world, pos, world.getBlockState(pos).getBlock().getDefaultState(), 1);
+            world.setBlockToAir(pos);
+            return false;
         }
         return true;
     }
 
-    public void mineBlocks(World world, ArrayList<BlockPos> miningBlocks, EntityCreature creature)
+    protected BlockPos getRandomBlock()
     {
-        while(!miningBlocks.isEmpty())
+        int x = 0;
+        int z = 0;
+        while(x == 0 && z == 0)
         {
-            BlockPos currentPosition = miningBlocks.get(0);
-            Block currentBlock = world.getBlockState(currentPosition).getBlock();
-            currentBlock.dropBlockAsItem(world, miningBlocks.get(0), currentBlock.getDefaultState(), 2);
-            world.setBlockToAir(currentPosition);
-            creature.getNavigator().tryMoveToXYZ(currentPosition.getX(), currentPosition.getY(), currentPosition.getZ(), 1);
-
-            miningBlocks.remove(0);
+            x = (int) (Math.random() * 3) - 1;
+            if(x == 0) z = (int) (Math.random() * 3) - 1;
         }
+        return entity.getPosition().add(x, 0, z);
+    }
+    protected Vec3d getPosition()
+    {
+        if (entity.isInWater())
+        {
+            Vec3d vec3d = RandomPositionGenerator.getLandPos(this.entity, 15, 7);
+            return vec3d == null ? RandomPositionGenerator.findRandomTarget(this.entity, 10, 7) : vec3d;
+        }
+        return this.entity.getRNG().nextFloat() >= this.probability ? RandomPositionGenerator.getLandPos(this.entity, 10, 7) : RandomPositionGenerator.findRandomTarget(this.entity, 10, 7);
     }
 }
